requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Flask==0.10.1
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
kombu==3.0.26
MarkupSafe==0.23
passlib==1.6.2
picamera==1.10
pygame==1.9.2a0
pyserial==2.7
pytz==2015.2
requests==2.7.0
RPi.GPIO==0.5.11
SQLAlchemy==1.0.4
Werkzeug==0.10.4
